Update - Tentatively completed all modules for personal design
Note that there is bound to be changes need for this, but mostly all modules in personal design is updated.

CPU is updated and configured to house all submodules, a few naming changes, and deleted unused signals from template.

Registers and ROM slightly changed also, ROM may need changing to accomodate for smaller instruction lengths