Saving before branch - trying to implement reduced instruction size
Noticed that the 8 bit immediate can be integrated with the src address of the instruction, should not interfere with other modules, but just need ot make sure that the op
type multiplexer in the top level (CPU.sv) works as intended.

If this fails, then go back to this save.