Completion - Just finished a working processor for affine transformation
Now the processor works for the affine transformation algorithm, and this is tested in affine_tb.sv. The results are stored in the top 2 gpr, and
contain the signed 2's complement answers. Note that the numbers are rounded down from the actual answer because you take the top 8 bits of the 16bit
multiplication.

Now we need to synthesise and put it onto an FPGA.