Working CPU with affine and synthesized on FPGA
Working CPU that correctly displays the affine transformation results on at least 2 tests. The synthesis shows no errors and is compiles with 94 ALMs;
this can be improved upon.

Firstly, I should reduce the number of registers to 3bit, therefor the instruction size is reduced to 12 bit in length, from 15. This will reduce the
numbers of registers, since many of them are not needed.