Currently working simulation of personal design v2 with port feature
Port module completed as a synchronous way of loading the switches into the correct registers. It also acts as a way to pause the program counter and display the results of X2 (switch
low) and Y2 (switch high).

Now to try on Quartus, although I doubt that I will have any less than 100 ALMs.